requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.20.0
    # via -r requirements.in
cattrs==22.1.0
    # via jsii
constructs==10.0.114
    # via
    #   -r requirements.in
    #   aws-cdk-lib
exceptiongroup==1.0.0rc3
    # via cattrs
jsii==1.56.0
    # via
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   constructs
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typing-extensions==4.1.1
    # via jsii
